+++ BALSAMIC/assets/scripts/collect_qc_metrics.py
@@ -8,8 +8,8 @@ import yaml
 import re
 
 from BALSAMIC.constants.metrics import METRICS
+from BALSAMIC.models.metrics import MetricModel
 from BALSAMIC.utils.io import read_json
-from BALSAMIC.utils.models import MetricModel
 from BALSAMIC.utils.rule import (
     get_capture_kit,
     get_sequencing_type,